chore(bigframes): move some ai accessor tests outside the SQLGlot compiler dir
